Visualise route networks with tmap

This function is for visualising route networks. It was originally developed to show quietness, as defined by CycleStreets

Usage

tm_rnet(
  rnet,
  palette = "-temperature_diverging",
  lwd_multiplier = 4,
  scale = 5,
  lwd = names(rnet)[1],
  col = names(rnet)[2],
  max_features = 10000,
  breaks = c(0, 25, 50, 75, 100),
  labels = c("0-24", "25-49", "50-75", "75+")
)

Arguments

rnet

palette

lwd_multiplier

The width of the largest line relative to the thinnest line

scale

The width of the widest line in the dataset

lwd

col

max_features

Value

A tmap object

Examples

tm_rnet(rnet_lisbon, lwd = "ENMAC4", col = "Quietness")
#> Error: object 'rnet_lisbon' not found
tm_rnet(rnet_lisbon, lwd = "ENMAC4", col = "Quietness", scale = 30)
#> Error: object 'rnet_lisbon' not found